Add config file path for different systems
I recently had to install this script on a Mac. Unfortunately, I kept running into issues with the login.

After some investigation, it turns out that `confy`, which is the package used to serialize and deserialize the `tempoit.toml` config, uses `directories-next` to locate the configuration file path.

According to the [`directories-next` docs](https://crates.io/crates/directories-next), the `config_dir`, which is `$HOME/.config` on Linux, is `$HOME/Library/Application Support` on Mac and `{FOLDERID_RoamingAppData}` on Windows.